Small changes (check for overwrites in log_arduino)

+++ log_arduino.py
@@ -1,11 +1,17 @@
 import time
 import serial
 import argparse
+import os.path
+import sys
 
 parser = argparse.ArgumentParser()
 parser.add_argument('-o', '--output', help='Output log file name')
 parser.add_argument('--port', help='address of serial port to read')
+parser.add_argument('-f', '--force', action='store_true', help='overwrite existing log file')
 args = parser.parse_args()
+if os.path.isfile(args.output) and not args.force:
+    print "Warning: file already exists. Run with -f to overwrite!"
+    sys.exit(0)
 fileout = open(args.output,'w')
 while True:
     # Weird serial read issue (duplicated across software)
@@ -16,4 +22,4 @@ while True:
             if len(output) > 0 and i != 0:
                 fileout.write(output)
                 print output
-            time.sleep(1)
+            time.sleep(0.5)
